requirements/prod.txt
python-dotenv==1.0.1
pyyaml==6.0.1
langchain==0.2.11
langchain-openai==0.1.19
tiktoken==0.7.0
streamlit==1.37.0
langchain-community==0.2.10
faiss-cpu==1.8.0
pypdf==4.3.1